+++ packages/external-clients/src/index.ts
@@ -1,1 +1,2 @@
 export * as ApimUtils from "./apim-client";
+export * as SelfcareUtils from "./selfcare-client";